# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger with Laravel",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/admin/comments": {
            "get": {
                "tags": [
                    "Comments"
                ],
                "summary": "Get all comments",
                "operationId": "faa69030aaa43aa3f648878d9f45d8ad",
                "responses": {
                    "200": {
                        "description": "All Comments"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/createComment": {
            "post": {
                "tags": [
                    "Comments"
                ],
                "summary": "Create a new comment post",
                "description": "Create a new comment post with the provided data",
                "operationId": "93bf8d6cd5bb5e81c021e1b273215a95",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "post_id",
                                    "user_id",
                                    "content"
                                ],
                                "properties": {
                                    "post_id": {
                                        "description": "The ID of the post the comment belongs to",
                                        "type": "integer",
                                        "format": "int64",
                                        "example": "123"
                                    },
                                    "user_id": {
                                        "description": "The ID of the user who posted the comment",
                                        "type": "integer",
                                        "format": "int64",
                                        "example": "456"
                                    },
                                    "content": {
                                        "description": "The content of the comment",
                                        "type": "string",
                                        "example": "This is a great post!"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successfully created a comment post",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "Indicates whether the request was successful",
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "description": "A message describing the outcome of the request",
                                            "type": "string",
                                            "example": "Created comment post successfully!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request. Invalid input data."
                    },
                    "401": {
                        "description": "Unauthorized. Authentication is required."
                    },
                    "500": {
                        "description": "Internal server error. Failed to create the comment post."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/comments/update/{id}": {
            "post": {
                "tags": [
                    "Comments"
                ],
                "summary": "Update a comment post by ID",
                "description": "Update a comment post by its ID",
                "operationId": "d00fd47499ba53fa16601b9f881e5541",
                "requestBody": {
                    "description": "Request body",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "content",
                                    "status"
                                ],
                                "properties": {
                                    "content": {
                                        "description": "Content of the comment post",
                                        "type": "string",
                                        "example": "Updated comment content"
                                    },
                                    "status": {
                                        "description": "Status of the comment post",
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "Validation error. Invalid input data."
                    },
                    "404": {
                        "description": "Not found. The comment post with the specified ID does not exist."
                    },
                    "500": {
                        "description": "Internal server error. Failed to update the comment post."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/deleteComment/{post_id}/{user_id}": {
            "delete": {
                "tags": [
                    "Comments"
                ],
                "summary": "Delete Comment Post",
                "operationId": "06378d166c7272feda28dc12ed7c9d5e",
                "parameters": [
                    {
                        "name": "post_id",
                        "in": "path",
                        "description": "Comment Post ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Comment not found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}